# docs/swagger.yaml
basePath: /api/v1
definitions:
  form.ToDoForm:
    properties:
      name:
        type: string
      username:
        type: string
    required:
    - name
    - username
    type: object
  form.ToDoResp:
    properties:
      id:
        type: string
      name:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  model.User:
    properties:
      id:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
host: localhost:8585
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Get all Todotask
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/form.ToDoResp'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all Todotask
    post:
      consumes:
      - application/json
      description: Create Todotask
      parameters:
      - description: ToDoForm
        in: body
        name: todoForm
        required: true
        schema:
          $ref: '#/definitions/form.ToDoForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/form.ToDoResp'
      summary: Create Todotask
  /todos/{id}:
    get:
      consumes:
      - application/json
      description: Get Todotask by id
      parameters:
      - description: ToDotask ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/form.ToDoResp'
      summary: Get Todotask by id
    put:
      consumes:
      - application/json
      description: Update Todotask
      parameters:
      - description: ToDoForm
        in: body
        name: todoForm
        required: true
        schema:
          $ref: '#/definitions/form.ToDoForm'
      - description: ToDotask ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/form.ToDoResp'
      summary: Update Todotask
  /user:
    get:
      consumes:
      - application/json
      description: Get all user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"